Split review.ts into focused modules
Keep server/lib/review.ts as the barrel exporting shared types and the
small core (package diff, risk computation, tar findings), and extract:

- review-rules (deterministicFindings + pattern arrays + rule IDs)
- review-diff-annotation (annotateFindingsWithDiffStatus + line diff)
- review-redaction (redactText/FileRecords/Findings/Json)
- review-package-files (package-files allowlist matchers)

firstMatchingLine now imports from text-utils. Pure refactor, no
behavior change.
