+++ server/lib/review.ts
@@ -1,6 +1,9 @@
-import { diffLines } from "diff";
-import { hasImplicitNodeGypInstall, isRootGypPath, normalizeStringRecord } from "./tar-parser.js";
 import type { TarSuspiciousEntry } from "./tar-parser.js";
+import {
+  DETERMINISTIC_RULE_IDS,
+  DETERMINISTIC_RULES_VERSION,
+  firstJsonPropertyLine,
+} from "./review-rules";
 
 export type RiskLevel = "low" | "medium" | "high" | "critical";
 
@@ -21,33 +24,6 @@ export interface Finding {
   ruleId?: string;
   ruleVersion?: string;
 }
-
-// Bump when deterministic rule semantics, severities, or coverage change in a
-// way that should invalidate cached scan reports. Stored alongside each
-// finding so historical reports can be traced back to the ruleset that
-// produced them.
-export const DETERMINISTIC_RULES_VERSION = "1.6.0";
-
-export const DETERMINISTIC_RULE_IDS = {
-  installScriptPreinstall: "install-script.preinstall",
-  installScript: "install-script.lifecycle",
-  codeProcessExecution: "code.process-execution",
-  codeNetworkAccess: "code.network-access",
-  codeDynamicEvaluation: "code.dynamic-evaluation",
-  codeCredentialAccess: "code.credential-access",
-  fileSecretContent: "file.secret-content",
-  fileLargeBinary: "file.large-binary",
-  fileNativeArtifact: "file.native-artifact",
-  fileOutsideFilesList: "file.outside-files-list",
-  installScriptImplicitNodeGyp: "install-script.implicit-node-gyp",
-  packageJsonParseFailed: "package-json.parse-failed",
-  diffCredentialFileAdded: "diff.credential-file-added",
-  diffLargeNewFile: "diff.large-new-file",
-  dependencyUnusualSpec: "dependency.unusual-spec",
-  dependencyOptionalAdded: "dependency.optional-added",
-  stageMetadataMismatch: "stage.metadata-mismatch",
-  tarSuspiciousEntry: "tar.suspicious-entry",
-} as const;
 
 export interface PackageJsonSummary {
   name?: string;
@@ -103,6 +79,8 @@ export interface FindingDiffAnnotation {
   releaseDelta: boolean;
 }
 
+export type CodePatternSet = "javascript" | "python";
+
 export interface FindingAnnotationOptions {
   previousFiles?: Array<Pick<FileRecord, "path" | "textSample" | "flags">>;
   stagedFiles?: Array<Pick<FileRecord, "path" | "textSample" | "flags">>;
@@ -110,327 +88,21 @@ export interface FindingAnnotationOptions {
   codePatternSet?: CodePatternSet;
 }
 
-export type CodePatternSet = "javascript" | "python";
+export {
+  DETERMINISTIC_RULE_IDS,
+  DETERMINISTIC_RULES_VERSION,
+  deterministicFindings,
+  PYTHON_EXECUTION_CAPABILITY_PATTERNS,
+} from "./review-rules";
+export type { DeterministicFindingOptions } from "./review-rules";
+export {
+  annotateFindingsWithDiffStatus,
+  isReleaseDeltaStatus,
+  normalizeFindingDiffStatus,
+} from "./review-diff-annotation";
+export { redactFileRecords, redactFindings, redactJson, redactText } from "./review-redaction";
 
-export interface DeterministicFindingOptions {
-  codePatternSet?: CodePatternSet;
-}
-
-const LIFECYCLE_SCRIPTS = ["preinstall", "install", "postinstall", "prepare"];
 const RISK_RANK: Record<RiskLevel, number> = { low: 0, medium: 1, high: 2, critical: 3 };
-const JS_PROCESS_EXECUTION_PATTERNS = [
-  /\bchild_process\b/,
-  /\bexecSync\b/,
-  /\bexecFileSync\b/,
-  /\bspawn\(/,
-  /\bspawnSync\(/,
-  /\bcurl\s/,
-  /\bwget\s/,
-  /\bnc\s/,
-  /\bbash\s+-c/,
-  /\bpowershell\s/,
-];
-const PYTHON_PROCESS_EXECUTION_PATTERNS = [
-  /\bsubprocess\b/,
-  /\bos\.system\s*\(/,
-  /\bos\.popen\s*\(/,
-  /\bPopen\s*\(/,
-  /\bpty\.spawn\s*\(/,
-  /\bcommands\.getoutput\s*\(/,
-];
-const JS_NETWORK_ACCESS_PATTERNS = [
-  /\brequire\(["'](?:node:)?(?:http|https|net|dns)["']\)/,
-  /\bfrom\s+["'](?:node:)?(?:http|https|net|dns)["']/,
-  /\bfetch\s*\(/,
-  /\bXMLHttpRequest\b/,
-  /\baxios\s*\./,
-];
-const PYTHON_NETWORK_ACCESS_PATTERNS = [
-  /\burllib\.request\b/,
-  /\brequests\.(?:get|post|put|patch|delete|request)\b/,
-  /\bhttp\.client\b/,
-  /\bhttplib\b/,
-  /\bsocket\.socket\s*\(/,
-  /\bftplib\b/,
-  /\bsmtplib\b/,
-  /\burlopen\s*\(/,
-];
-const JS_DYNAMIC_EVALUATION_PATTERNS = [
-  /\beval\s*\(/,
-  /\bnew\s+Function\s*\(/,
-  /\bWebAssembly\.compile\s*\(/,
-  /\batob\s*\(/,
-  /\bBuffer\.from\s*\([^,]+,\s*["']base64["']\s*\)/,
-];
-const PYTHON_DYNAMIC_EVALUATION_PATTERNS = [
-  /(?<!\.)\bexec\s*\(/,
-  /\b__import__\s*\(/,
-  /\bimportlib\.import_module\s*\(/,
-  /\bmarshal\.loads\s*\(/,
-  /(?<!\.)\bcompile\s*\(/,
-  /\bbase64\.b(?:64|32|16)decode\s*\(/,
-  /\bzlib\.decompress\s*\(/,
-  /\blzma\.decompress\s*\(/,
-  /\bcodecs\.decode\s*\(/,
-  /\bbytes\.fromhex\s*\(/,
-];
-const JS_CREDENTIAL_ACCESS_PATTERNS = [
-  /\bprocess\.env\b/,
-  /\bnpm_config_/,
-  /\bNPM_TOKEN\b/,
-  /\bGITHUB_TOKEN\b/,
-  /\bAWS_SECRET\b/,
-  /\bPRIVATE_KEY\b/,
-];
-const PYTHON_CREDENTIAL_ACCESS_PATTERNS = [
-  /\bos\.environ\b/,
-  /\bos\.getenv\s*\(/,
-  /\bgetpass\b/,
-  /\bkeyring\b/,
-  /\.aws\/credentials/,
-  /\.ssh\/id_/,
-  /\.netrc/,
-];
-
-const JS_PATTERN_SET = {
-  processExecution: JS_PROCESS_EXECUTION_PATTERNS,
-  networkAccess: JS_NETWORK_ACCESS_PATTERNS,
-  dynamicEvaluation: JS_DYNAMIC_EVALUATION_PATTERNS,
-  credentialAccess: JS_CREDENTIAL_ACCESS_PATTERNS,
-};
-const PYTHON_PATTERN_SET = {
-  processExecution: PYTHON_PROCESS_EXECUTION_PATTERNS,
-  networkAccess: PYTHON_NETWORK_ACCESS_PATTERNS,
-  dynamicEvaluation: PYTHON_DYNAMIC_EVALUATION_PATTERNS,
-  credentialAccess: PYTHON_CREDENTIAL_ACCESS_PATTERNS,
-};
-
-// Python process-execution, network, and dynamic-evaluation capability in one set.
-// Reused by ecosystem adapters that need to know whether a file executes code
-// (e.g. a PyPI sdist's setup.py, which pip runs at install time).
-export const PYTHON_EXECUTION_CAPABILITY_PATTERNS = [
-  ...PYTHON_PROCESS_EXECUTION_PATTERNS,
-  ...PYTHON_NETWORK_ACCESS_PATTERNS,
-  ...PYTHON_DYNAMIC_EVALUATION_PATTERNS,
-];
-
-const SECRET_PATTERNS: Array<[RegExp, string]> = [
-  [/npm_[A-Za-z0-9]{20,}/g, "[REDACTED_NPM_TOKEN]"],
-  [/gh[pousr]_[A-Za-z0-9_]{20,}/g, "[REDACTED_GITHUB_TOKEN]"],
-  [/github_pat_[A-Za-z0-9_]{20,}/g, "[REDACTED_GITHUB_TOKEN]"],
-  [/AKIA[0-9A-Z]{16}/g, "[REDACTED_AWS_ACCESS_KEY]"],
-  [/ASIA[0-9A-Z]{16}/g, "[REDACTED_AWS_SESSION_KEY]"],
-  [/AIza[0-9A-Za-z\-_]{35}/g, "[REDACTED_GOOGLE_API_KEY]"],
-  [/ya29\.[0-9A-Za-z\-_]{20,}/g, "[REDACTED_GOOGLE_OAUTH_TOKEN]"],
-  [/sk_(?:live|test)_[0-9a-zA-Z]{16,}/g, "[REDACTED_STRIPE_KEY]"],
-  [/rk_(?:live|test)_[0-9a-zA-Z]{16,}/g, "[REDACTED_STRIPE_KEY]"],
-  [/xox[abprs]-[A-Za-z0-9-]{10,}/g, "[REDACTED_SLACK_TOKEN]"],
-  [/https:\/\/hooks\.slack\.com\/services\/[A-Za-z0-9/]+/g, "[REDACTED_SLACK_WEBHOOK]"],
-  [/eyJ[A-Za-z0-9_-]{10,}\.[A-Za-z0-9_-]{10,}\.[A-Za-z0-9_-]{10,}/g, "[REDACTED_JWT]"],
-  [/\b(?:[A-Za-z]+:\/\/)[^\s/@:]+:[^\s/@]+@[^\s'"\\]+/g, "[REDACTED_URL_WITH_CREDENTIALS]"],
-  [
-    /-----BEGIN (?:RSA |OPENSSH |EC |DSA |PGP |ENCRYPTED )?PRIVATE KEY(?: BLOCK)?-----[\s\S]*?-----END (?:RSA |OPENSSH |EC |DSA |PGP |ENCRYPTED )?PRIVATE KEY(?: BLOCK)?-----/g,
-    "[REDACTED_PRIVATE_KEY]",
-  ],
-  [/(authorization\s*[:=]\s*)['"]?Bearer\s+[A-Za-z0-9._\-+/=]{16,}/gi, "$1[REDACTED_BEARER]"],
-  [
-    /((?:secret|token|password|passwd|pwd|api[_-]?key|access[_-]?key|client[_-]?secret)\s*[:=]\s*)['"]?[^'"\s]{12,}/gi,
-    "$1[REDACTED_SECRET]",
-  ],
-];
-
-export function deterministicFindings(
-  files: FileRecord[],
-  diff: DiffEntry[] = [],
-  packageJsonSummary?: PackageJsonSummary | null,
-  options: DeterministicFindingOptions = {},
-): Finding[] {
-  const findings: Finding[] = [];
-  const patterns = codePatternsFor(options.codePatternSet);
-  const diffByPath = new Map(diff.map((entry) => [entry.path, entry]));
-  const packageJsonFile = files.find((file) => file.path === "package.json" && file.textSample);
-  const rawPackageJson = packageJsonFile?.textSample
-    ? (safeJson(packageJsonFile.textSample) as PackageJsonSummary | null)
-    : null;
-  const packageJsonParseFailed = Boolean(packageJsonFile?.textSample) && rawPackageJson === null;
-  const packageJson = packageJsonSummary ?? rawPackageJson;
-  const scripts = normalizeStringRecord(packageJson?.scripts);
-  const implicitScripts = normalizeStringRecord(packageJson?.implicitScripts);
-  const rootGypFile = files.find((file) => isRootGypPath(file.path));
-  const tag = (
-    rule: keyof typeof DETERMINISTIC_RULE_IDS,
-    finding: Omit<Finding, "ruleId" | "ruleVersion">,
-  ): Finding => ({
-    ...finding,
-    ruleId: DETERMINISTIC_RULE_IDS[rule],
-    ruleVersion: DETERMINISTIC_RULES_VERSION,
-  });
-
-  if (packageJsonParseFailed) {
-    findings.push(
-      tag("packageJsonParseFailed", {
-        severity: "medium",
-        file: packageJsonFile?.path ?? "package.json",
-        line: 1,
-        evidence: packageJsonFile?.flags.includes("truncated")
-          ? "package.json parse failed; captured sample was truncated"
-          : "package.json parse failed",
-        reason:
-          "the package manifest could not be parsed, so lifecycle script and dependency review from the tarball manifest may be incomplete",
-      }),
-    );
-  }
-
-  const implicitNodeGyp =
-    implicitScripts.install === "node-gyp rebuild" || hasImplicitNodeGypInstall(files, packageJson);
-  if (implicitNodeGyp) {
-    findings.push(
-      tag("installScriptImplicitNodeGyp", {
-        severity: "high",
-        file: rootGypFile?.path ?? packageJsonFile?.path ?? "package.json",
-        line: rootGypFile ? 1 : firstJsonPropertyLine(packageJsonFile?.textSample, "gypfile"),
-        evidence: "implicit install: node-gyp rebuild",
-        reason: rootGypFile
-          ? "npm defaults install to node-gyp rebuild when a root *.gyp file exists and no install/preinstall script or gypfile=false is declared"
-          : "npm staged metadata reports the default node-gyp install hook; the source root had a *.gyp file even if that file is not present in the packed tarball",
-      }),
-    );
-  }
-
-  for (const script of LIFECYCLE_SCRIPTS) {
-    if (!scripts[script] || implicitScripts[script] === scripts[script]) continue;
-    findings.push(
-      tag(script === "preinstall" ? "installScriptPreinstall" : "installScript", {
-        severity: script === "preinstall" ? "critical" : "high",
-        file: packageJsonFile?.path ?? "package.json",
-        line: firstJsonPropertyLine(packageJsonFile?.textSample, script, scripts[script]),
-        evidence: `${script}: ${scripts[script]}`,
-        reason: "install lifecycle hooks execute on consumer machines",
-      }),
-    );
-  }
-
-  for (const file of files) {
-    const sample = file.textSample || "";
-    const changed = diffByPath.get(file.path)?.status;
-    const changedPrefix = changed && changed !== "unchanged" ? `new/changed ${changed} file: ` : "";
-
-    if (patterns.processExecution.some((pattern) => pattern.test(sample))) {
-      findings.push(
-        tag("codeProcessExecution", {
-          severity: "high",
-          file: file.path,
-          line: firstMatchingLine(sample, patterns.processExecution),
-          evidence: `${changedPrefix}process or shell execution`,
-          reason: "package may execute arbitrary commands",
-        }),
-      );
-    }
-    if (patterns.networkAccess.some((pattern) => pattern.test(sample))) {
-      findings.push(
-        tag("codeNetworkAccess", {
-          severity: changed === "added" ? "high" : "medium",
-          file: file.path,
-          line: firstMatchingLine(sample, patterns.networkAccess),
-          evidence: `${changedPrefix}network-capable code path`,
-          reason:
-            "unexpected network access in package code can be used for exfiltration or staged payload retrieval",
-        }),
-      );
-    }
-    if (patterns.dynamicEvaluation.some((pattern) => pattern.test(sample))) {
-      findings.push(
-        tag("codeDynamicEvaluation", {
-          severity: changed === "added" ? "high" : "medium",
-          file: file.path,
-          line: firstMatchingLine(sample, patterns.dynamicEvaluation),
-          evidence: `${changedPrefix}dynamic code or obfuscation primitive`,
-          reason: "common malware and obfuscation technique",
-        }),
-      );
-    }
-    if (patterns.credentialAccess.some((pattern) => pattern.test(sample))) {
-      findings.push(
-        tag("codeCredentialAccess", {
-          severity: changed === "added" ? "high" : "medium",
-          file: file.path,
-          line: firstMatchingLine(sample, patterns.credentialAccess),
-          evidence: `${changedPrefix}secret/environment access`,
-          reason: "package may read credentials from the install environment",
-        }),
-      );
-    }
-    if (/\.npmrc|\.env|id_rsa|id_ed25519/i.test(file.path) || containsSecretLikeText(sample)) {
-      findings.push(
-        tag("fileSecretContent", {
-          severity: changed === "added" ? "critical" : "high",
-          file: file.path,
-          line: firstSecretLine(sample),
-          evidence: `${changedPrefix}secret-looking file or content`,
-          reason: "published artifacts should not include credentials or private material",
-        }),
-      );
-    }
-    if (file.flags.includes("binary") && file.size > 1024 * 1024) {
-      findings.push(
-        tag("fileLargeBinary", {
-          severity: changed === "added" ? "high" : "info",
-          file: file.path,
-          evidence: `${file.size} byte binary`,
-          reason: "large binary should be reviewed manually",
-        }),
-      );
-    }
-    if (isOutsidePackageFilesAllowlist(file.path, packageJson) && changed !== "removed") {
-      findings.push(
-        tag("fileOutsideFilesList", {
-          severity: changed === "added" ? "high" : "medium",
-          file: file.path,
-          evidence: `${changedPrefix}file is not matched by package.json files allowlist`,
-          reason:
-            "unexpected files outside the declared package files list can indicate tarball tampering or generated payloads that are not visible in the source/package manifest review",
-        }),
-      );
-    }
-    if (/\.(node|dll|so|dylib|exe|wasm)$/i.test(file.path)) {
-      findings.push(
-        tag("fileNativeArtifact", {
-          severity: "high",
-          file: file.path,
-          evidence: "native, wasm, or executable artifact",
-          reason:
-            "native binaries are hard to audit and can execute outside JavaScript policy checks",
-        }),
-      );
-    }
-  }
-
-  for (const entry of diff) {
-    if (entry.status === "added" && /(^|\/)(\.npmrc|\.env|id_rsa|id_ed25519)$/i.test(entry.path)) {
-      findings.push(
-        tag("diffCredentialFileAdded", {
-          severity: "critical",
-          file: entry.path,
-          line: 1,
-          evidence: "credential-looking file added",
-          reason: "package artifact includes a file name commonly associated with secrets",
-        }),
-      );
-    }
-    if (entry.status === "added" && entry.stagedSize && entry.stagedSize > 2 * 1024 * 1024) {
-      findings.push(
-        tag("diffLargeNewFile", {
-          severity: "medium",
-          file: entry.path,
-          evidence: `${entry.stagedSize} byte new file`,
-          reason: "large new package artifact should be reviewed",
-        }),
-      );
-    }
-  }
-
-  return findings;
-}
 
 export function tarSuspiciousEntryFindings(
   entries: TarSuspiciousEntry[] | undefined | null,
@@ -589,214 +261,6 @@ export function computeRisk(findings: Array<{ severity?: string | null }>): Risk
   return "low";
 }
 
-export function annotateFindingsWithDiffStatus<
-  T extends { id?: string; file: string; line?: number | null; ruleId?: string | null },
->(
-  findings: T[],
-  diff: Array<{ path: string; status?: unknown }>,
-  options: FindingAnnotationOptions = {},
-): Array<T & FindingDiffAnnotation> {
-  const diffByPath = new Map(
-    diff.map((entry) => [entry.path, normalizeFindingDiffStatus(entry.status)]),
-  );
-  const previousByPath = new Map((options.previousFiles ?? []).map((file) => [file.path, file]));
-  const stagedByPath = new Map((options.stagedFiles ?? []).map((file) => [file.path, file]));
-  const changedLineCache = new Map<string, Set<number> | null>();
-  return findings.map((finding) => {
-    const persisted = finding.id ? options.persistedAnnotations?.get(finding.id) : null;
-    if (persisted) return { ...finding, ...persisted };
-
-    const diffStatus = diffByPath.get(finding.file) ?? "unknown";
-    return {
-      ...finding,
-      diffStatus,
-      releaseDelta:
-        isReleaseScopedFinding(finding) ||
-        isNewlyEnabledImplicitNodeGypFinding(finding, previousByPath, stagedByPath) ||
-        isFindingOnReleaseDelta(
-          finding,
-          diffStatus,
-          previousByPath,
-          stagedByPath,
-          changedLineCache,
-          options.codePatternSet,
-        ),
-    };
-  });
-}
-
-function isReleaseScopedFinding(finding: { ruleId?: string | null }): boolean {
-  return Boolean(
-    finding.ruleId?.startsWith("stage.") ||
-    finding.ruleId?.startsWith("pypi.") ||
-    finding.ruleId === DETERMINISTIC_RULE_IDS.dependencyUnusualSpec ||
-    finding.ruleId === DETERMINISTIC_RULE_IDS.dependencyOptionalAdded ||
-    finding.ruleId === DETERMINISTIC_RULE_IDS.diffCredentialFileAdded ||
-    finding.ruleId === DETERMINISTIC_RULE_IDS.diffLargeNewFile ||
-    finding.ruleId === DETERMINISTIC_RULE_IDS.tarSuspiciousEntry,
-  );
-}
-
-function isNewlyEnabledImplicitNodeGypFinding(
-  finding: { ruleId?: string | null },
-  previousByPath: Map<string, Pick<FileRecord, "path" | "textSample" | "flags">>,
-  stagedByPath: Map<string, Pick<FileRecord, "path" | "textSample" | "flags">>,
-): boolean {
-  if (finding.ruleId !== DETERMINISTIC_RULE_IDS.installScriptImplicitNodeGyp) return false;
-
-  const stagedPackageJson = parsePackageJsonFile(stagedByPath.get("package.json"));
-  if (!hasImplicitNodeGypInstall([...stagedByPath.values()], stagedPackageJson)) return false;
-
-  const previousPackageJson = parsePackageJsonFile(previousByPath.get("package.json"));
-  return !hasImplicitNodeGypInstall([...previousByPath.values()], previousPackageJson);
-}
-
-function parsePackageJsonFile(
-  file: Pick<FileRecord, "textSample"> | undefined,
-): PackageJsonSummary | null {
-  if (!file?.textSample) return null;
-  return safeJson(file.textSample) as PackageJsonSummary | null;
-}
-
-function isFindingOnReleaseDelta(
-  finding: { file: string; line?: number | null; ruleId?: string | null },
-  diffStatus: FindingDiffStatus,
-  previousByPath: Map<string, Pick<FileRecord, "path" | "textSample" | "flags">>,
-  stagedByPath: Map<string, Pick<FileRecord, "path" | "textSample" | "flags">>,
-  changedLineCache: Map<string, Set<number> | null>,
-  codePatternSet: CodePatternSet | undefined,
-): boolean {
-  if (diffStatus === "added") return true;
-  if (diffStatus !== "modified") return false;
-  if (!finding.line) return true;
-
-  const changedLines = changedStagedLinesForPath(
-    finding.file,
-    previousByPath,
-    stagedByPath,
-    changedLineCache,
-  );
-  if (!changedLines) return true;
-  if (changedLines.has(finding.line)) return true;
-  return findingPatternMatchesChangedLine(
-    finding,
-    stagedByPath.get(finding.file)?.textSample,
-    changedLines,
-    codePatternSet,
-  );
-}
-
-function changedStagedLinesForPath(
-  path: string,
-  previousByPath: Map<string, Pick<FileRecord, "path" | "textSample" | "flags">>,
-  stagedByPath: Map<string, Pick<FileRecord, "path" | "textSample" | "flags">>,
-  cache: Map<string, Set<number> | null>,
-): Set<number> | null {
-  if (cache.has(path)) return cache.get(path) ?? null;
-  const previous = previousByPath.get(path);
-  const staged = stagedByPath.get(path);
-  if (!previous?.textSample || !staged?.textSample) {
-    cache.set(path, null);
-    return null;
-  }
-  if (previous.flags.includes("binary") || staged.flags.includes("binary")) {
-    cache.set(path, null);
-    return null;
-  }
-  const lines = changedStagedLines(previous.textSample, staged.textSample);
-  cache.set(path, lines);
-  return lines;
-}
-
-function changedStagedLines(previous: string, staged: string): Set<number> {
-  const changed = new Set<number>();
-  let stagedLine = 0;
-  for (const part of diffLines(previous, staged)) {
-    const lines = splitComparableLines(part.value);
-    if (part.added) {
-      for (const _line of lines) {
-        stagedLine += 1;
-        changed.add(stagedLine);
-      }
-    } else if (!part.removed) {
-      stagedLine += lines.length;
-    }
-  }
-  return changed;
-}
-
-function splitComparableLines(text: string): string[] {
-  const lines = text.split("\n");
-  if (lines.length && lines[lines.length - 1] === "") lines.pop();
-  return lines;
-}
-
-function findingPatternMatchesChangedLine(
-  finding: { file: string; ruleId?: string | null },
-  stagedText: string | undefined,
-  changedLines: Set<number>,
-  codePatternSet: CodePatternSet | undefined,
-): boolean {
-  if (!stagedText) return false;
-  const patterns = patternsForFinding(finding, codePatternSet);
-  if (!patterns.length) return false;
-  const lines = splitComparableLines(stagedText);
-  for (const lineNumber of changedLines) {
-    const line = lines[lineNumber - 1];
-    if (line === undefined) continue;
-    for (const pattern of patterns) {
-      pattern.lastIndex = 0;
-      if (pattern.test(line)) return true;
-    }
-  }
-  return false;
-}
-
-function patternsForFinding(
-  finding: { file: string; ruleId?: string | null },
-  codePatternSet: CodePatternSet | undefined,
-): RegExp[] {
-  const patterns = codePatternSet
-    ? codePatternsFor(codePatternSet)
-    : finding.file.endsWith(".py")
-      ? PYTHON_PATTERN_SET
-      : JS_PATTERN_SET;
-  switch (finding.ruleId) {
-    case DETERMINISTIC_RULE_IDS.codeProcessExecution:
-      return patterns.processExecution;
-    case DETERMINISTIC_RULE_IDS.codeNetworkAccess:
-      return patterns.networkAccess;
-    case DETERMINISTIC_RULE_IDS.codeDynamicEvaluation:
-      return patterns.dynamicEvaluation;
-    case DETERMINISTIC_RULE_IDS.codeCredentialAccess:
-      return patterns.credentialAccess;
-    case DETERMINISTIC_RULE_IDS.fileSecretContent:
-      return SECRET_PATTERNS.map(([pattern]) => pattern);
-    default:
-      return [];
-  }
-}
-
-function codePatternsFor(codePatternSet: CodePatternSet | undefined): typeof JS_PATTERN_SET {
-  return codePatternSet === "python" ? PYTHON_PATTERN_SET : JS_PATTERN_SET;
-}
-
-export function isReleaseDeltaStatus(status: FindingDiffStatus): boolean {
-  return status === "added" || status === "modified";
-}
-
-export function normalizeFindingDiffStatus(value: unknown): FindingDiffStatus {
-  switch (value) {
-    case "added":
-    case "removed":
-    case "modified":
-    case "unchanged":
-      return value;
-    default:
-      return "unknown";
-  }
-}
-
 export function combineRisk(...risks: Array<RiskLevel | null | undefined>): RiskLevel {
   return risks.reduce<RiskLevel>((highest, risk) => {
     if (!risk) return highest;
@@ -808,134 +272,6 @@ export function normalizeRisk(value: unknown): RiskLevel {
   return value === "critical" || value === "high" || value === "medium" || value === "low"
     ? value
     : "medium";
-}
-
-export function redactText(text: string): string {
-  return SECRET_PATTERNS.reduce(
-    (out, [pattern, replacement]) => out.replace(pattern, replacement),
-    text,
-  );
-}
-
-function containsSecretLikeText(text: string): boolean {
-  return SECRET_PATTERNS.some(([pattern]) => {
-    pattern.lastIndex = 0;
-    return pattern.test(text);
-  });
-}
-
-function firstMatchingLine(
-  text: string | undefined | null,
-  patterns: RegExp[],
-): number | undefined {
-  if (!text) return undefined;
-  const lines = text.split("\n");
-  for (let index = 0; index < lines.length; index += 1) {
-    for (const pattern of patterns) {
-      pattern.lastIndex = 0;
-      if (pattern.test(lines[index])) return index + 1;
-    }
-  }
-  return undefined;
-}
-
-function firstSecretLine(text: string | undefined | null): number | undefined {
-  if (!text) return undefined;
-  return firstMatchingLine(
-    text,
-    SECRET_PATTERNS.map(([pattern]) => pattern),
-  );
-}
-
-function firstJsonPropertyLine(
-  text: string | undefined | null,
-  key: string,
-  value?: string,
-): number | undefined {
-  if (!text) return undefined;
-  const escapedKey = escapeRegExp(key);
-  const keyPattern = new RegExp(`["']${escapedKey}["']\\s*:`);
-  const lines = text.split("\n");
-  for (let index = 0; index < lines.length; index += 1) {
-    if (keyPattern.test(lines[index])) return index + 1;
-  }
-  if (value) {
-    const escapedValue = escapeRegExp(value);
-    const valuePattern = new RegExp(escapedValue);
-    for (let index = 0; index < lines.length; index += 1) {
-      if (valuePattern.test(lines[index])) return index + 1;
-    }
-  }
-  return undefined;
-}
-
-export function redactFileRecords(files: FileRecord[]): FileRecord[] {
-  return files.map((file) => ({
-    ...file,
-    textSample: file.textSample ? redactText(file.textSample) : file.textSample,
-  }));
-}
-
-export function redactFindings(findings: Finding[]): Finding[] {
-  return findings.map((finding) => ({
-    ...finding,
-    evidence: redactText(finding.evidence),
-    reason: redactText(finding.reason),
-  }));
-}
-
-export function redactJson<T>(value: T): T {
-  if (typeof value === "string") return redactText(value) as T;
-  if (Array.isArray(value)) return value.map((item) => redactJson(item)) as T;
-  if (value && typeof value === "object") {
-    return Object.fromEntries(
-      Object.entries(value as Record<string, unknown>).map(([key, nested]) => [
-        key,
-        redactJson(nested),
-      ]),
-    ) as T;
-  }
-  return value;
-}
-
-function isOutsidePackageFilesAllowlist(
-  path: string,
-  packageJson: PackageJsonSummary | null | undefined,
-): boolean {
-  const files = packageJson?.files?.filter((entry) => entry.trim()) ?? [];
-  if (!files.length) return false;
-  if (isAlwaysIncludedPackageFile(path, packageJson)) return false;
-  return !files.some((entry) => packageFilesEntryMatches(entry, path));
-}
-
-function isAlwaysIncludedPackageFile(
-  path: string,
-  packageJson: PackageJsonSummary | null | undefined,
-): boolean {
-  const lower = path.toLowerCase();
-  if (lower === "package.json") return true;
-  if (/^(?:readme|licen[cs]e|copying|notice)(?:\.|$)/i.test(path)) return true;
-  const entrypoints = [packageJson?.main, packageJson?.module, packageJson?.types];
-  if (entrypoints.includes(path)) return true;
-  const bin = packageJson?.bin;
-  if (typeof bin === "string" && bin === path) return true;
-  if (bin && typeof bin === "object" && Object.values(bin).includes(path)) return true;
-  return false;
-}
-
-function packageFilesEntryMatches(entry: string, path: string): boolean {
-  const normalized = entry.trim().replace(/^\.\//, "").replace(/\/+$/, "");
-  if (!normalized) return false;
-  if (normalized.includes("*")) return globLikePackageFilesEntryMatches(normalized, path);
-  return path === normalized || path.startsWith(`${normalized}/`);
-}
-
-function globLikePackageFilesEntryMatches(entry: string, path: string): boolean {
-  const escaped = entry
-    .replace(/[.+?^${}()|[\]\\]/g, "\\$&")
-    .replace(/\*\*/g, ".*")
-    .replace(/\*/g, "[^/]*");
-  return new RegExp(`^${escaped}$`).test(path);
 }
 
 function diffDependencySections(
@@ -978,16 +314,4 @@ function diffObject(
       out.push({ key, status: "modified", previous: before[key], staged: after[key] });
   }
   return out;
-}
-
-function escapeRegExp(value: string): string {
-  return value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-function safeJson(text: string): unknown | null {
-  try {
-    return JSON.parse(text);
-  } catch {
-    return null;
-  }
 }
